web_app/requirements.txt
flask==3.0.2
pillow==10.2.0
numpy==1.24.3
torch==2.1.2
torchvision==0.16.2
gunicorn==21.2.0 